Cache system controller software list for prestage orchestration
Previously, the system controller software list was requested once for
each subcloud during prestage orchestration. This was unnecessary and
could overload the software API, which operates with a single worker.
This commit leverages the existing DC orchestrator's in-memory cache
solution to also cache the system controller software list, similar to
how the software orchestrator already handles it.

Test Plan:
01. PASS - Run a subcloud prestage orchestration for software
    deployment with multiple subclouds and verify that it works as
    expected, ensuring that only a single software list request is
    made to the system controller software API.
02. PASS - Perform the same test using prestage for install.
03. PASS - Run a single subcloud prestage operation using the
    "dcmanager subcloud prestage" command, for both software deployment
    and install. Verify that it still works as expected.

Partial-Bug: 2092319

Change-Id: I7ff7728afdc1864fa04285deb6d556ec73967c81
Signed-off-by: Gustavo Herzmann <gustavo.herzmann@windriver.com>