Improve available releases deletion logic
In the finish_strategy step of the Software Deploy Orchestration,
available releases are deleted only if they are not deployed in the
System Controller. This change ensures that prestaged releases are
preserved, while still cleaning up the releases on the remove process.

Test Plan:
PASS: Execute a sw-deploy-orchestration without errors
PASS: Prestage releases and apply a previous release
- SystemController has .0, .100, and .200 releases deployed
- Prestage releases to the subcloud
- Apply .100 to the subcloud using sw-deploy-strategy
- Verify that the .200 release remains available and is not deleted
PASS: Prestage releases, apply and remove.
- After prestage, remove .200 from SystemController
- Apply .100 to the subcloud using sw-deploy-strategy
- Verify that the .200 release is deleted from the subcloud

Closes-bug: 2101156

Change-Id: I923978c06276d2d470d1061e34016ae95ee99be2
Signed-off-by: Hugo Brito <hugo.brito@windriver.com>