Add dedicated queue for state change notifications
This change introduces a dedicated queue for auditing subclouds
triggered by state change notifications from dcmanager-state
(e.g., subcloud online). The new queue is processed by a separate
periodic task that runs more frequently to ensure timely handling
of these events.

The main reason for introducing this new queue is to properly handle
re-audit timing in case of failures. With a single queue where online
subclouds are enqueued with priority 0, it's not possible to delay
the retry by 60 seconds, as the timestamp is used as the priority
to control processing order. As a result, the subcloud would be
placed at the end of the queue, and it may take some time until it
is audited again.

A lock per subcloud is used to avoid concurrent audits of the same
subcloud from both queues.

The logging string format was corrected to align with the rest of
the logs in Distributed Cloud.

The unit tests for dccertmon will be added in a separate commit.

Test Plan:
PASS: Start dccertmon in an environment where many subclouds are
      turned off, causing audits to timeout. Then, turn on some
      subclouds and verify that they are audited immediately.
PASS: Restart the dccertmon service while subclouds are coming online.
      Verify that as both queues are populated, there are no
      concurrent audits of the same subcloud, and no lock usage
      errors.
PASS: Change the admin password and ensure that there is no need of
      updating the dccertmon keystone password since it's using the
      sysinv credentials.

Story: 2011311
Task: 51961

Change-Id: I5d8d53dff2fbd2bd48f1bbb075a9bbd3059e9ba4
Signed-off-by: Enzo Candotti <Enzo.Candotti@windriver.com>