Fix releases not being removed in sw-deploy-strategy
After commit [1], sw-deploy-strategy allows available releases
to exist in the subcloud if they're deployed in the SystemController.
This causes an issue in pre_check when the user tries to remove a
release but still has available releases prestaged. This commit
fixes that by checking if the release the user wants to apply is
the latest deployed in the subcloud, even if there are available
releases that need to be deleted.

1: https://review.opendev.org/c/starlingx/distcloud/+/943640

Test Plan:
PASS: Execute a sw-deploy-orchestration without errors
PASS: Prestage releases and apply a previous release
- SystemController has .0, .100, and .200 releases deployed
- Prestage releases to the subcloud
- Apply .100 to the subcloud using sw-deploy-strategy
- Verify that the .200 release remains available and is not deleted
PASS: Prestage releases, apply and remove.
- After prestage, remove .200 from SystemController
- Apply .100 to the subcloud using sw-deploy-strategy
- Verify that the .200 release is deleted from the subcloud
PASS: Prestage releases and apply.
- Remove .200 from subcloud (.200 is available and not deleted)
- Remove .100 and .200 releases from SystemController
- Apply .0 to the subcloud using sw-deploy-strategy
- Verify that the .100 is removed; .100 and .200 releases are
  deleted from the subcloud in the end
PASS: Release already removed.
- SystemController has .0, .100 deployed
- Subcloud has .0, .100 deployed and .200 available
  (force subcloud to be out-of-sync based on deploy_status)
- Apply .100 to the subcloud using sw-deploy-strategy
- Orchestration will skip directly to finish_strategy,
  .200 release deleted and subcloud in-sync.

Closes-bug: 2106763

Change-Id: I77108212f111e155de1a4e25dff07b729a62eda5
Signed-off-by: Hugo Brito <hugo.brito@windriver.com>