Allow passing a list of excluded operations to the rvmc script
This commit adds logic to specify a list of predefined
operations to exclude from the BMC sequence.
This allows the BMC sequence to be modified via the
--exclude_operations parameter. If the parameter is
not specified, the default sequence will be executed.

This commit relates to the enrollment functionality,
where the subcloud reboot should be skipped when the
seed iso is detected.

Test Plan:

PASS: - Complete the factory install on the subcloud.
      - Run the enrollment process through dcmanager and
        verify that it completes successfully.
      - In the first stage of enrollment, when the seed
        ISO is inserted, verify that the subcloud does
        not reboot. Enrollment should complete without
        the need to reboot the subcloud.

Regression Test:
      - New subcloud deployment (subcloud add)
      - Subcloud restore (or redeploy) post enrollment

Depends-On: https://review.opendev.org/c/starlingx/utilities/+/947441

Closes-bug: 2109795

Change-Id: I33c9dd8dd0df37adec1cb3cb9ac619d3115b99f9
Signed-off-by: Cristian Mondo <cristian.mondo@windriver.com>