Mask sensitive information in dcorch logging
This commit masks sensitive information, such as passwords,
from being exposed in dcorch logs. When a password is present
in the log data, it is replaced with asterisks for security.

Additionally, `synchronize_session="fetch"` is added to SQLAlchemy
API to prevent `InvalidRequestError` when updating records after a
`sync_request`.

Test Plan:
PASS: Change a user's password and confirm that logs mask the password.
- Execute: openstack --os-region-name SystemController user password set
- dcorch logs show: 'password': '******'

Closes-bug: 2109374

Change-Id: I042e1592b5bbd4f996f5604a54234235965ed0b3
Signed-off-by: Hugo Brito <hugo.brito@windriver.com>