Introduce auto-restore parameters to dcmanager API
This commit adds support for the 'auto' and 'factory' auto-restore
parameters to the dcmanager API. It also includes preliminary support
for these new parameters within the dcmanager services.

Currently, the API rejects the use of these parameters until the
full auto-restore feature is implemented.

Test Plan:
Temporarily remove the pecan.abort call when using 'auto' and
'factory', and run the following tests:
01. PASS - Run a subcloud restore with --auto and verify that
    gen-bootloader-iso.sh is called with the --auto-restore-config
    parameter and that the install playbook is called with the
    auto_restore_mode extra var.
02. PASS - Repeat the previous test, but use --factory instead.
03. PASS - Run a subcloud restore with --auto and --factory and verify
    that the generated restore_values yml file contains the
    auto_restore_mode field.
04. PASS - Run a subcloud restore with both --auto and --factory at
    the same time and verify that --factory takes precedence.
05. PASS - Modify the code to fake a subcloud install success and
    verify that the restore playbook is not called when using --auto
    or --factory.

Add the pecan.abort call again and run the following tests:
06. PASS - Run a subcloud restore with --auto and --factory with a
    subcloud that doesn't have install_values and check that the
    operation is aborted with the correct error message.
07. PASS - Run a subcloud restore with --auto and/or --factory and
    verify that the request is aborted.

Note that the auto-restore feature is not yet implemented, so the tests
only validate up to the point where the install/auto-restore operation
would be invoked.

Story: 2011454
Task: 52110

Change-Id: I76173f1c4e1f315edd980bafab53d03ff51b09ed
Signed-off-by: Gustavo Herzmann <gustavo.herzmann@windriver.com>