Cleanup dcmanager audit
This removes the old dcmanager audit that was performed without dcagent
and updates the unit tests. It additionally removes the openstack audit
that was unused.

Note: the audit's unit tests were removed because they test the
deleted methods, requiring additional changes to ensure accurate code
execution, which will be done in a further commit.

Test plan:
1. PASS: Successfully run linters.
2. PASS: Manage a subcloud and verify that all endpoints are properly
   set to in-sync.
3. PASS: Unmanage a subcloud and verify that all endpoints are properly
   set to unknown.

Story: 2011311
Task: 52158

Change-Id: I8acd4cfbac45b003afdd9833835126afac30fbc0
Signed-off-by: Raphael <Raphael.Lima@windriver.com>