Avoid passing --auto-restore-config when auto restore is disabled
Prevent passing the overrides_file argument to _run_subcloud_install()
during a subcloud restore with install when auto or factory restore
is not in use. Without this change, --auto-restore-config would be
propagated to gen-bootloader-iso.sh, which doesn't support it.

Test Plan:
Temporarily remove the pecan.abort call when using 'auto' and
'factory', and run the following tests:
01. PASS - Run a subcloud restore with --auto or --factory and verify
    that gen-bootloader-iso.sh is invoked with --auto-restore-config.
02. PASS - Run a subcloud restore with --with-install but without
    --auto or --factory and verify that gen-bootloader-iso.sh is
    invoked without --auto-restore-config.

Story: 2011454
Task: 52110

Change-Id: Ie985f596bbe466618deda98b8c53af52ff19003d
Signed-off-by: Gustavo Herzmann <gustavo.herzmann@windriver.com>