Update dcmanager API to support cloud-init config
This commit supports the dcmanager API to accept optional cloud-init
config, verify if it is a tarball and store it in dc-vault. It is
expected later to be used to generate the seed ISO for subcloud
enrollment, and the tarball need to be removed after that.

Test plan:
1. Passed - add subcloud with enroll and cloud-init-config tarball,
verified tarball stored to dc-vault correctly.
2. Passed - deploy enroll subcloud with cloud-init-config tarball,
verified tarball updated in dc-vault.

Story: 2011455
Task: 52372

Signed-off-by: Yuxing Jiang <Yuxing.Jiang@windriver.com>
Change-Id: I695db1560531c4519fa2636fee76d39eac18379a