Fix endpoint cache group in dccertmon
This commit adds the missing configuration options to the
endpoint_cache group for dccertmon. Without this change, the dccertmon
service fails to start due to a NoSuchOptError.

Test Plan:
PASS: Run an upgrade from N-1 to current version with these changes.
      Verify that the dccertmon service is running correctly, and the
      swact can be done.
PASS: Apply this change in a master load. Verify that the service can
      be started with no errors.

Closes-Bug: 2116278

Change-Id: Ie1becdbb9b8f6e109bea994bd970b1d3822d20f7
Signed-off-by: Enzo Candotti <Enzo.Candotti@windriver.com>