Prevent auto-restore if subcloud is not aio-simplex
This commit retrieves the system mode of a subcloud by inspecting the
bootstrap values file stored in the dc-vault, and prevents the
auto-restore operation from proceeding if the subcloud is not simplex.

It also fixes an unrelated unit test for the install prep phase, which
was taking 60 seconds to execute due to the lack of mocking for the
"urlretrieve" method, causing it to wait until timeout.

Test Plan:
01. PASS - Modify the system mode in the subcloud bootstrap values to
    "duplex" and execute an auto-restore. Verify that the API aborts
    the request with the appropriate error message.
02. PASS - Repeat the previous test using a factory restore instead.
03. PASS - Run a normal restore and verify that the API does not abort
    the request.
04. PASS - Change the system mode back to "simplex" and attempt both
    auto and factory restore. Verify that the API does not abort the
    requests.

Story: 2011454
Task: 52536

Change-Id: I1ee3095a84815484a2f668d174530d53f8c0cd83
Signed-off-by: Gustavo Herzmann <gustavo.herzmann@windriver.com>