Remove the logic that determines whether the release is a base release
Currently, when the for-install prestage is run and
a release is passed through the command line, it
checks if that release is in the software list API,
validating whether the listed release is the initial
release (base release). Whose mechanism to determine
if it is a base release is by checking if the patch
part of the version is zero (MM.pp.0), which is wrong,
preventing a release that has its patch part in the
version other than zero from being prestaged, for
example the release MM.mm.1.

This commit removes the logic that validates whether
the patch part is 0 in the release version, allowing
the provided release to be validated in any of the
three allowed states for the prestage: deployed,
available, and unavailable.

In this way, any release that is in any of those
three states within the software list can be verified
with the release provided on the command line through
the MM.mm part.

Test Plan:
PASS - Bring up a SystemController and a subcloud with
       master load.
     - Upload release 24.09 to the SystemController.
     - Prestage the subcloud for-install providing
       release 24.09
     - Verify that the release is prestaged correctly.

Closes-bug: 2117290

Change-Id: Ie74bc9cb9af5a69a0572952d9f4c9d44f63c18a4
Signed-off-by: Cristian Mondo <cristian.mondo@windriver.com>