Caching software deploy state result
Implemented a cache system for the result for
is_valid_software_deploy_state to improve process
of adding and rehoming multiple subclouds. The
result of the function call is cached for 1
minute, this is valid because the result of this
function shouldn't change that often. The cache
system stores the result and replicates it,
doing that in a synchronized way to prevent
conflicts.

Test plan:
  - PASS: Add 250 subclouds until create step and
          verify no errors occurred and the
          software api calls were reduced

Story: 2011311
Task: 52583

Change-Id: Ia2aa717e0457f2b8a6b54f69daea4aaa9ea88aef
Signed-off-by: Giovane Ribeiro Anhaia <giovane.ribeiroanhaia@windriver.com>