Rework IPMI SEL monitoring to remove delays during enrollment
This commit updates the IPMI SEL monitoring script and the
99-completion-event enrollment script to eliminate the required delays.
The monitoring script now supports a parameter to retrieve the last SEL
event ID and another to specify the initial ID to use during
monitoring.

With these changes, the playbook can determine the initial ID before
mounting the seed ISO and pass it to the monitoring task, removing the
need to sleep between each enrollment stage.

Test Plan:
01. PASS - Run subcloud enrollment and verify that the custom
    cloud-init complete IPMI SEL event is sent and that the enroll-init
    completes successfully.
02. PASS - Add a failing custom script before 99-completion-event,
    verify the completion script does not execute due to
    run-parts --exit-on-error.
03. PASS - Set ipmi_sel_event_monitoring to false inside the install
    values and run the subcloud enrollment. Verify that the
    99-completion-event script is not created.
04. PASS - Run subcloud enrollment and verify that the last event is
    successfully obtained, and the consequent events are received
    correctly.

Story: 2011455
Task: 52656

Change-Id: I14917188623dc93ceefa3a77cdd1ee4f8b17ca9d
Signed-off-by: Gustavo Herzmann <gustavo.herzmann@windriver.com>