Disable ipmi monitoring without cloud-init tarball
The IPMI sel event monitoring takes time to query the sel event from
the BMC controller. It is not necessary to monitor the sel event in
the case if no additional cloud-init configuration offered. Monitoring
the sel event by default may cause additional time consumption not
expected.

This commit checks if the cloud-init config tarball existed in
dc-vault, disables the IPMI monitoring if it was not specified in
the subcloud's install values.

Test plan:
1. Verified the tasks for monitoring the IPMI events are skipped if no
   cloud-init config and ipmi_sel_event_monitoring was not specified
   in the install values.
2. Verified the tasks for monitoring the IPMI events are not skipped
   if cloud-init config are appended and ipmi_sel_event_monitoring was
   not specified in the install values.
3. Verified the tasks for monitoring the IPMI events are not skipped
   if cloud-init config are not appended and ipmi_sel_event_monitoring
   was True in the install values.

Closes-bug: 2122444
Signed-off-by: Jiang, Yuxing <Yuxing.Jiang@windriver.com>
Change-Id: Ia106f5581475bd78294a2fa327c1828019da90eb