Sync user's passwd_last_change field
Password expiration relies on the last changed date to calculate the
actual expiration. Previously, only the password hash and max age were
synchronized between hosts. If the last changed date differed between
hosts, having only the max age would still result in inconsistent
expiration dates.

This change adds support for a new field, passwd_last_change, when
synchronizing user's password with subclouds. It will be ignored if the
subcloud does not support it.

Test Plan:
PASS: build-pkgs & build-image.
PASS: In an AIO-DX, first install the controller-0, move the clock of
      controller-1 forward a few days before installing it. After
      finishing bootstrap, check that the "Password expires" field is
      the same for both controllers when running `chage -l sysadmin`.
PASS: In an AIO-DX, add a worker node and check the password expiration
      as above.
PASS: In a DC, add a subcloud days after central cloud installation and
      check the password expiration is the same as central's.
PASS: Upgrade an AIO-DX with unsynchronized password expiration and
      verify they are synchronized after upgrade.
PASS: In a DC setup, with the central cloud running this patch and an
      AIO-SX subcloud still on the old version, confirm backward
      compatibility: passwords remain synchronized even though
      `passwd_last_change` is ignored on the subcloud.

Closes-Bug: 2122715
Depends-On: https://review.opendev.org/c/starlingx/config/+/959340

Change-Id: Ie41d60f494d4e6403d70dc366ef8e21e51672541
Signed-off-by: Rodrigo Tavares <Rodrigo.DosSantosTavares@windriver.com>