Allow USM prestage for-sw-deploy for N-1 subcloud
Currently the prestage is restricted to only allow prestage for
releases in deployed state. The problem is that for N-1 releases
which are in unavailable state, cannot be prestaged.

This change removes the state restriction for N-1 software
versions in the prestage for-sw-deploy workflow, allowing them
to be prestaged even if they are in unavailable state.
For N releases the state restriction is still applied, that is
the release must be in deployed state to be prestaged.
This change also removes the restriction for N-1 releases in the
creating VIM strategy strategy workflow, allowing them to be
deployed even if they are in unavailable state.

Additional changes:
  - Remove snapshot parameter from sw_deploy strategy workflow.
    The snapshot parameter was intruduced as part of rollback
    support in review:
    https://review.opendev.org/c/starlingx/distcloud/+/956307
    This parameter should be removed from VIM API payload because
    the N-1 software does not support snapshot. If the subcloud
    has a N-1 release and the snapshot parameter is not supported,
    the VIM API will return an error.

Test Plan:
PASS - Prestage for-sw-deploy with N-1 release in Unavailable state
     - Verify that the N-1 release is prestaged in the Subcloud and
       state is Available.
     - Run the sw-deploy strategy workflow with the N-1 release and
       verify that it deployed successfully.

PASS - Create and apply prestage-strategy for-sw-deploy with N-1
       release in Unavailable state
     - Verify that the N-1 release is prestaged in the Subcloud and
       state is Available.
     - Run the sw-deploy strategy workflow with the N-1 release and
       verify that it deployed successfully.

Regression test:
     - Validate sw-deploy strategy with N release and snapshot.
     - Apply the release then rollback.
     - Create sw_deploy strategy providing snapshot parameter with
       N release
     - Prestage N release on N-1 subcloud for deployment
     - Prestage N release on N subcloud for deployment
     - Prestage N release on N-1 subcloud for install
     - Prestage N-1 release on N-1 subcloud for install
     - Prestage N release on N subcloud for install
     - Prestage N-1 release on N subcloud for install

Depends-On: https://review.opendev.org/c/starlingx/update/+/961834

Closes-Bug: 2125230

Change-Id: I8d48c597a85e063c585a433adc8d5d513b0a5cd6
Signed-off-by: Cristian Mondo <cristian.mondo@windriver.com>