Add timeout to retrieve subcloud region name while rehoming
Currently, when rehoming a subcloud, the system
attemps to retrieve the region name of the subcloud
to register it with the System Controller.
In a scale environment the response may take a
long time if the subcloud are offline, causing
the rehome operation to stuck the environment.
To prevent this, a timeout is added to the region
name retrieval operation, allowing the rehome to
proceed even if the region name is not retrieved
in a timely manner.

Test Plan:
PASS - Prepare a scale environment with 250
       subclouds.
     - Power off all subclouds.
     - Rehome the subclouds in parallel.
     - Verify the rehome operation completes
       successfully and the environment is stable.
     - Verify the response time for subcloud show
       command is within acceptable limits.
PASS - Same as above but with online subclouds.

Closes-Bug: 2125696

Change-Id: Id9940036e689c587f063101f5f86c59d86ddc82e
Signed-off-by: Cristian Mondo <cristian.mondo@windriver.com>