Add rollback validation for sw-deploy-strategy
This commit introduces a new validation in sw-deploy-strategy,
pre_check state. The validation checks if the subcloud has any
software releases in the `deploying` state, which makes it a valid
candidate for rollback.

If no releases are in the `deploying` state, the rollback
operation is not required, and the orchestration is completed
immediately with deploy_status `complete`.

Test Plan:
PASS: Peform a sw-deploy-strategy successfully.
PASS: Attempt a rollback when a subcloud has a release in the
      `deploying` state and verify the rollback proceeds.
PASS: Attempt a rollback when no releases are in the
      `deploying` state and verify the orchestration completes
      successfully with status `complete`.
PASS: Fail if a sw-upgrade strategy already exists with a proper
      message. (Perform a rollback if a release is deploying)

Closes-bug: 2125789

Change-Id: I84ce8dd9eb568846a9dff6b742b2757fcf87ab26
Signed-off-by: Hugo Brito <hugo.brito@windriver.com>