Send system controller timestamp to enroll-init
After a factory install, the system can be powered off for extended
periods of time, making the subcloud enrollment to fail. To fix this
issue, the time on the servers will need to be updated during the
subcloud enrollment.

This commit updates the subcloud enrollment code to send the system
controller timestamp as a parameter of the enroll-init-reconfigure,
which will be used for updating the subcloud time.

By default the enroll-timestamp parameter is not sent, unless the
'update_subcloud_clock' override is set to True in the subcloud's
install values.

In addition,the create_enroll_override_file function was updated to
fix formatting errors that occurred when new enroll_overrides values
were added.

Test Plan:
1. PASS: Verified that the enroll-timestamp parameter is not sent to
   the 10-platform-reconfig script if update_subcloud_clock was not
   specified in the install values.
2. PASS: Verified that the enroll-timestamp parameter is not sent to
   the 10-platform-reconfig script if update_subcloud_clock was set to
   'False' in the install values.
3. PASS: Verified that the enroll-timestamp parameter is sent to the
   10-platform-reconfig script if update_subcloud_clock was set to
   'True' in the install values.

Depends-On: https://review.opendev.org/c/starlingx/utilities/+/963229

Closes-Bug: https://bugs.launchpad.net/starlingx/+bug/2126942

Change-Id: I0595050cba9179bea66259e108c9a08d8d29abc1
Signed-off-by: Enzo Candotti <Enzo.Candotti@windriver.com>