Fix oam floating ip retrieval in prestage strategy
This commit fixes the oam floating ip retrieval in prestage strategy by
recovering the information from the subcloud in case it is not stored in
the worker's memory. Previously, on a service restart or host swact, the
data would be lost, causing the step to fail with a KeyError.

Test plan:
1. PASS: Create a prestage strategy and apply it successfully.
2. PASS: Create a prestage strategy, apply it and restart the service
   after the steps are moved out of pre-check. Verify that the workers
   are able to recover the subcloud's oam ip and continue the strategy
   application.
3. PASS: Run test 2 in scale.

Closes-Bug: 2127105

Change-Id: Icc66434781e779a98db40acc36afba31107ffd05
Signed-off-by: Raphael <Raphael.Lima@windriver.com>